move path template handling under settings
